test(ssh): split lost-worker coverage into its own file
