+++ src/main/ssh/ssh-relay-session.test.ts
@@ -6,15 +6,9 @@ import {
   AGENT_HOOK_INSTALL_PLUGINS_METHOD
 } from '../../shared/agent-hook-relay'
 import { SSH_RELAY_CONFIGURE_GRACE_TIME_METHOD } from '../../shared/ssh-types'
-import { getDefaultWorkspaceSession } from '../../shared/constants'
-import type { RelayPtyLostEntry } from '../../shared/pty-revive-protocol'
-import type { WorkspaceSessionState } from '../../shared/types'
 import { createMockDeps, mockDeploySuccess } from './ssh-relay-session-test-fixtures'
 
-const { muxRequestMock, archiveLostTerminalWorkerMock } = vi.hoisted(() => ({
-  muxRequestMock: vi.fn(),
-  archiveLostTerminalWorkerMock: vi.fn()
-}))
+const { muxRequestMock } = vi.hoisted(() => ({ muxRequestMock: vi.fn() }))
 
 vi.mock('./ssh-relay-deploy', () => ({
   deployAndLaunchRelay: vi.fn()
@@ -22,10 +16,6 @@ vi.mock('./ssh-relay-deploy', () => ({
 
 vi.mock('./ssh-relay-deploy-helpers', () => ({
   execCommand: vi.fn().mockResolvedValue('')
-}))
-
-vi.mock('../terminal-lost-worker-archive', () => ({
-  archiveLostTerminalWorker: archiveLostTerminalWorkerMock
 }))
 
 vi.mock('./ssh-channel-multiplexer', () => {
@@ -120,76 +110,6 @@ const { registerSshGitProvider, unregisterSshGitProvider } =
 const { routeExternalPtyData, routeExternalPtyReplay, routeExternalPtyExit } =
   await import('../ipc/pty-renderer-delivery-router')
 
-const REVIVE_WORKTREE_ID = 'repo-1::/worktree'
-const REVIVE_TAB_ID = 'tab-1'
-const REVIVE_LEAF_ID = '11111111-1111-4111-8111-111111111111'
-const REVIVE_PANE_KEY = `${REVIVE_TAB_ID}:${REVIVE_LEAF_ID}`
-const RELAY_LOST_WORKER: RelayPtyLostEntry = {
-  id: 'ssh:target-1@@pty-lost',
-  kind: 'recognized-worker',
-  reason: 'process-not-running',
-  pid: 42,
-  cols: 80,
-  rows: 24,
-  cwd: '/repo',
-  worktreeId: REVIVE_WORKTREE_ID,
-  tabId: REVIVE_TAB_ID,
-  paneKey: REVIVE_PANE_KEY
-}
-
-function workspaceSessionForRelayLostWorker(workerHint = false): WorkspaceSessionState {
-  return {
-    ...getDefaultWorkspaceSession(),
-    tabsByWorktree: {
-      [REVIVE_WORKTREE_ID]: [
-        {
-          id: REVIVE_TAB_ID,
-          worktreeId: REVIVE_WORKTREE_ID,
-          title: 'Worker',
-          customTitle: null,
-          color: null,
-          sortOrder: 0,
-          createdAt: 10,
-          ptyId: RELAY_LOST_WORKER.id
-        }
-      ]
-    },
-    terminalLayoutsByTabId: {
-      [REVIVE_TAB_ID]: {
-        root: { type: 'leaf', leafId: REVIVE_LEAF_ID },
-        activeLeafId: REVIVE_LEAF_ID,
-        expandedLeafId: null
-      }
-    },
-    terminalPtyIncarnationsByPaneKey: { [REVIVE_PANE_KEY]: 'incarnation-1' },
-    terminalArchiveHintsByPaneKey: {
-      [REVIVE_PANE_KEY]: workerHint ? { launchAgent: 'codex', startedAt: 10 } : { cwd: '/repo' }
-    }
-  }
-}
-
-function configureStagedPtyRevive(args: {
-  serialize: ReturnType<typeof vi.fn>
-  revive: ReturnType<typeof vi.fn>
-  reattachPtyIds?: string[]
-}): { attachForReconnect: ReturnType<typeof vi.fn> } {
-  const attachForReconnect = vi.fn().mockResolvedValue({})
-  vi.mocked(getSshPtyProvider)
-    .mockReset()
-    .mockReturnValueOnce({
-      serialize: args.serialize,
-      dispose: vi.fn()
-    } as unknown as ReturnType<typeof getSshPtyProvider>)
-    .mockReturnValue({
-      revive: args.revive,
-      attachForReconnect,
-      dispose: vi.fn()
-    } as unknown as ReturnType<typeof getSshPtyProvider>)
-  vi.mocked(getPtyIdsForConnection).mockReturnValueOnce(['ssh:target-1@@pty-lost'])
-  vi.mocked(getPtyIdsForConnection).mockReturnValue(args.reattachPtyIds ?? [])
-  return { attachForReconnect }
-}
-
 function mockReconnectPtyProvider(attachForReconnect: ReturnType<typeof vi.fn>): void {
   vi.mocked(getSshPtyProvider).mockReturnValue({
     attachForReconnect,
@@ -215,7 +135,6 @@ describe('SshRelaySession', () => {
     delete process.env.ORCA_FEATURE_REMOTE_AGENT_HOOKS
     muxRequestMock.mockReset()
     muxRequestMock.mockResolvedValue([])
-    archiveLostTerminalWorkerMock.mockReset()
     vi.mocked(getSshPtyProvider)
       .mockReset()
       .mockReturnValue({
@@ -513,108 +432,6 @@ describe('SshRelaySession', () => {
 
     expect(mockAttach).toHaveBeenCalledWith('pty-1')
     expect(mockAttach).toHaveBeenCalledWith('pty-2')
-  })
-
-  it.each([
-    'typed recognized',
-    'typed unclassified worker hint',
-    'typed unclassified',
-    'legacy',
-    'malformed typed',
-    'serialize failure',
-    'archive failure'
-  ])('keeps staged PTY revive %s on its compatible recovery path', async (state) => {
-    const { mockConn, mockStore, session } = await establishRelaySession()
-    vi.clearAllMocks()
-    mockDeploySuccess()
-    const serialize = vi.fn().mockResolvedValue('staged-pty-state')
-    const revive = vi.fn()
-    const lost = state.startsWith('typed unclassified')
-      ? { ...RELAY_LOST_WORKER, kind: 'unclassified' as const }
-      : RELAY_LOST_WORKER
-    if (state === 'serialize failure') {
-      serialize.mockRejectedValue(new Error('old relay stopped before serialize reply'))
-    } else if (state === 'malformed typed') {
-      revive.mockRejectedValue(new Error('typed outcome validation failed'))
-    } else if (state === 'legacy') {
-      revive.mockResolvedValue({
-        mode: 'legacy',
-        diagnosticCode: 'pty-revive-outcome-unavailable',
-        outcome: { outcomeVersion: 1, revived: [], lost: [RELAY_LOST_WORKER], diagnostics: [] }
-      })
-    } else {
-      revive.mockResolvedValue({
-        mode: 'typed',
-        outcome: { outcomeVersion: 1, revived: [], lost: [lost], diagnostics: [] }
-      })
-    }
-    archiveLostTerminalWorkerMock.mockResolvedValue(
-      state === 'archive failure'
-        ? { kind: 'error', code: 'durability-failed' }
-        : {
-            kind: 'archived',
-            archive: {},
-            operationId: 'relay-worker-lost:tab-1',
-            ptyIdsToKill: [RELAY_LOST_WORKER.id]
-          }
-    )
-    mockStore.getWorkspaceSession = vi
-      .fn()
-      .mockReturnValue(
-        workspaceSessionForRelayLostWorker(state === 'typed unclassified worker hint')
-      )
-    const archived = state === 'typed recognized' || state === 'typed unclassified worker hint'
-    const fallbackPtyIds = archived ? [] : ['ssh:target-1@@pty-fallback']
-    const { attachForReconnect } = configureStagedPtyRevive({
-      serialize,
-      revive,
-      reattachPtyIds: fallbackPtyIds
-    })
-
-    await session.reconnect(mockConn)
-
-    expect(serialize).toHaveBeenCalledWith(['pty-lost'], { formatVersion: 2 })
-    expect(revive).toHaveBeenCalledTimes(state === 'serialize failure' ? 0 : 1)
-    expect(archiveLostTerminalWorkerMock).toHaveBeenCalledTimes(
-      archived || state === 'archive failure' ? 1 : 0
-    )
-    expect(mockStore.markSshRemotePtyLease).toHaveBeenCalledTimes(
-      archived ? 1 : fallbackPtyIds.length
-    )
-    expect(mockStore.markSshRemotePtyLease).toHaveBeenCalledWith(
-      'target-1',
-      archived ? 'pty-lost' : 'pty-fallback',
-      archived ? 'expired' : 'attached'
-    )
-    expect(routeExternalPtyExit).toHaveBeenCalledTimes(archived ? 1 : 0)
-    expect(session.getState()).toBe('ready')
-    if (state === 'typed unclassified') {
-      expect(mockStore.getWorkspaceSession).toHaveBeenCalledWith('ssh:target-1')
-    }
-    if (state === 'typed unclassified worker hint') {
-      expect(archiveLostTerminalWorkerMock).toHaveBeenCalledWith(
-        expect.objectContaining({
-          candidate: expect.objectContaining({
-            relayEvidence: expect.objectContaining({ kind: 'unclassified' })
-          })
-        })
-      )
-    }
-    if (fallbackPtyIds.length > 0) {
-      expect(attachForReconnect).toHaveBeenCalledWith('pty-fallback')
-    }
-    if (state === 'archive failure') {
-      expect(archiveLostTerminalWorkerMock).toHaveBeenCalledWith(
-        expect.objectContaining({
-          candidate: expect.objectContaining({ relayEvidence: RELAY_LOST_WORKER })
-        })
-      )
-      expect(mockStore.markSshRemotePtyLease).not.toHaveBeenCalledWith(
-        'target-1',
-        'pty-lost',
-        'expired'
-      )
-    }
   })
 
   it('forwards reconnect replay after the attach attempt is still current', async () => {
